refactor: split 클래스 병합
